requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.32,<1
pymongo[srv]>=4.6,<5
pydantic-settings>=2.6,<3
boto3>=1.35,<2

# dev / tests
pytest>=8.0,<9
httpx>=0.27,<1
